Add option root build path
Many environments are pretty limited on disk space, especially the /tmp
partition could be quite small. For our builds on macOS we got a large
partition mounted to /Users/Shared/drone which we would like to use for
the Drone builds, this flag/env variable brings in the option to move
the build root to a different base folder.